feat(charts): add JSON schema validation for multiple charts
Added JSON schema files for the following charts to ensure proper validation of values:
- cyberchef
- dnscrypt-proxy
- mktxp
- nfs-server
- technitium-dnsserver
- winbox

Additionally, updated the `servicemonitor.yaml` template in the mktxp chart to include proper setup of variables and merging of labels and annotations. Also, modified the `values.yaml` for the winbox chart to set `KASM_DEBUG` as a string value for consistency. These changes enhance the configuration management and validation process across the charts.