chore(charts): update version and appVersion to 1.3.1 and 13.2.2
Updated the Chart.yaml and README.md files to reflect the new version and application version of the Technitium DNS Server chart. The version has been bumped from 1.3.0 to 1.3.1, and the appVersion has been updated from 13.1 to 13.2.2. Additionally, the change type in the annotation has been modified to reflect these updates.