chore(opengist): update helm chart version and documentation
Updated the Chart.yaml to increment the version from 1.0.2 to 1.0.3 and added new data such as sources and maintainers. Improved the description formatting for better readability. Enhanced README.md with additional instructions for Helm repository usage, installation, and configuration of the chart. Created a new NOTES.txt to provide post-deployment guidance, and detailed values.yaml with comments for each configuration option to improve usability and understanding of the Helm chart configurations.